Clear the checkpoint before the backfill's orphan refusal
An orphan refusal ended the walk leaving the checkpoint row behind. An
operator repairs an orphan by adding legs that sort before the walked
cursor, so a retried walk resumed past the repaired region and reported
the same orphan forever. Clear the checkpoint first, then refuse, so a
repaired re-run walks from the start. The regression test pins the
cleared cursor after a refusal.

Also split the refunded projection suite into select-refunded/ files,
each under the 400-line test limit, with the support file's failure arms
pinned directly so its guards stay fully covered. Issue #2363 tracks the
late-leg fence a review asked for, which needs the maintenance-task
subsystem.
